requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
agno>=1.0.0
httpx>=0.27.0
rich>=13.9.0
typer>=0.12.0
python-dotenv>=1.0.0
psutil>=6.0.0
sse-starlette>=2.1.0
ollama>=0.4.0
